Add "I" (int) to the list of "common file types"
 I just got badly bit trying to use PIL on a 16-bit png.  I was hunting desperately in docs and the built-in help
 for the list of "common file types" and could
not find it.  I suspected that my problem had to do with opening as RGB or "L".  This is arguably a (documented) bug
in PIL.  If you don't want to include my parenthetical expression explaining what happens if you get the filetype
wrong, I am entirely OK with that, but please accept the inclusion of "I" in your documentation on "common filetypes".
Here is my parenthetical expression -- and you can see it in my code:  (It should be noted that opening a 16-bit image in “L” or “RGB” will result in all pixels above 255 being set to 255.)  Thanks all!